🐛 유저 라우터 버그 수정

+++ src/routers/user-router.js
@@ -11,13 +11,14 @@ const userRouter = Router();
 
 userRouter.get('/', loginRequired, async (req, res, next) => {
 	try {
-		const { userId } = req.currentUserId;
+		const user = await userService.getUserByReq(req);
+		const userId = user._id;
 
 		if (!userId) {
 			throw new Error('userId 값이 없습니다.');
 		}
-		const user = await userService.getUser(userId);
-		res.status(200).json(user);
+		const userInfo = await userService.getUser(userId);
+		res.status(200).json(userInfo);
 	} catch (err) {
 		next(err);
 	}
@@ -28,10 +29,17 @@ userRouter.patch(
 	loginRequired,
 	changeNicknameValidator,
 	async (req, res, next) => {
-		const { userId } = req.currentUserId;
+
 		const { newNickname } = req.body;
 
 		try {
+			const user = await userService.getUserByReq(req);
+			const userId = user._id;
+
+			if (!userId) {
+				throw new Error('userId 값이 없습니다.');
+			}
+
 			const updateNickname = await userService.editUserNickname(
 				userId,
 				newNickname,
@@ -50,10 +58,17 @@ userRouter.patch(
 	loginRequired,
 	addFoodValidator,
 	async (req, res, next) => {
-		const { userId } = req.currentUserId;
+
 		const { addFoodId } = req.body;
 
 		try {
+			const user = await userService.getUserByReq(req);
+			const userId = user._id;
+
+			if (!userId) {
+				throw new Error('userId 값이 없습니다.');
+			}
+
 			const updateUserFood = await userService.addUserFood(userId, addFoodId);
 			res.status(200).json(updateUserFood);
 		} catch (err) {
@@ -64,7 +79,12 @@ userRouter.patch(
 
 userRouter.delete('/', loginRequired, async (req, res, next) => {
 	try {
-		const { userId } = req.currentUserId;
+		const user = await userService.getUserByReq(req);
+		const userId = user._id;
+
+		if (!userId) {
+			throw new Error('userId 값이 없습니다.');
+		}
 		const deletedUser = await userService.deleteUser(userId);
 
 		res.status(204).json(deletedUser);
